New Improvements & features
External modulations are much more faster now.
Instead of calculating the variables on the fly, all the variables are pre-calculated.

Phase of the ERASynth can be changed now.
Phase control is added to firware and both GUI.

In WEB GUI, frequency and amplitude increment/decrement steps are added.

When reference is external, OCXO and TCXO are turned off.
When reference is internal, only selected source is turned on.

Firmware Version : v1.0.16
WEB GUI Version : v1.1.6
ESP firmware version : v.1.0.6